// food-channel-processor/src/v1/grpc/interfaces/customer-order-for-channel-food.ts
export interface FoodInComboResponse {
  id: number;
  name: string;
  price: string;
  quantity: string;
}

export interface FoodInAdditionResponse {
  id: number;
  name: string;
  price_with_temporary: string;
  quantity: string;
}

export interface CustomerOrderOnlineDetailResponse {
  id: number;
  customer_order_id: number;
  food_id: number;
  food_name: string;
  order_detail_parent_id: number;
  order_detail_combo_parent_id: number;
  customer_order_detail_addition_ids: string;
  customer_order_detail_combo_ids: string;
  quantity: string;
  price: string;
  total_price: string;
  is_addition: number;
  is_combo: number;
  note: string;
  customer_order_detail_combo: FoodInComboResponse[];
  customer_order_detail_addition: FoodInAdditionResponse[];
  total_price_addition: string;
  created_at: string;
  updated_at: string;
}